+++ test/dactyl_keyboard/param/schema/parse_test.clj
@@ -43,12 +43,12 @@
 (deftest compass-compatible-angle-parser
   (testing "string"
     (is (= (parse/compass-compatible-angle "N") 0.0))
-    (is (= (parse/compass-compatible-angle "NE") (/ π 4)))
-    (is (= (parse/compass-compatible-angle "west") (* 1.5 π)))
+    (is (= (parse/compass-compatible-angle "NE") (/ π -4)))
+    (is (= (parse/compass-compatible-angle "west") (* -1.5 π)))
     (is (= (parse/compass-compatible-angle "π * 3") (* 3 π))))
   (testing "keyword"
     (is (= (parse/compass-compatible-angle :N) 0.0))
-    (is (= (parse/compass-compatible-angle :NE) (/ π 4))))
+    (is (= (parse/compass-compatible-angle :NE) (/ π -4))))
   (testing "number"
     (is (= (parse/compass-compatible-angle 0) 0))
     (is (= (parse/compass-compatible-angle 0.1) 0.1)))